Critical fixes to form stub lifetime management
BLUF: Several parts of the UI didn't properly clear uses, by working-copy data and similar data, of to-be-deleted forms. The vast majority of these have been fixed; the Quest editing dialog still needs auditing.

DovahKit features loaded form data that is, by default, "managed:" forms refer to each other via `form_reference_t` objects which function as tracked pointers, with setters that keep use info up-to-date bidirectionally. Several parts of the UI, however, need to be able to create and retain their own copies of form-data structs. A sensible design would be to have "managed" and "unmanaged" versions of the form-data types (and any sub-object types therein), with the UI using the latter; metaprogramming could be used to implement conversion between the two. However, I was a lot less skilled at metaprogramming, and a lot less sensible, when I built DovahKit's foundations half a decade ago. So instead, we have the ability to create a single "working copy" of a form's loaded data. In essence, "managed" and "unmanaged" form data use the same underlying class, and the setters on `form_reference_t` just politely decline to update Use Info when fed a "working copy" form.

This is incredibly limiting; among other things, it means you can only overwrite whole forms, bulldozing them in their entirety. (This is problematic for things like the list of event conditions on quests: edited via the SM*N UI, but the QUST UI has to take care not to clobber it.) It also blurs the line between "managed" and "unmanaged," such that when writing UI code, it's easy to forget that you need to sever uses of to-be-deleted forms yourself. But these two problems -- working with form data wholesale, and making it easy to forget that *you are the manager now* -- actually synergize into an even worse problem, because usually, UIs don't need a *full* copy of the *entire* form.

For simple form types, like Shout, all of the form's constituent data can be represented in the states of the UI's individual widgets: plop six DKFormPickers in a dialog and Bob's your uncle. For more complex form types, like Quest, not all form data is exposed through unique, individual widgets in the UI all at once; the form-editing dialog for a Quest will only display one quest stage at a time, for example, and so the data for quest stages can't live in the UI widgets; this is, indeed, why "working copies" were deemed necessary. Ideally, we'd be able to pick and choose where each individual piece of data lives -- in UI-owned unmanaged form data, or in a single widget -- but because the "working copy" system isn't granular enough, we don't get to pick. Instead, state has to be manually copied out of the "working copy" and into those widgets in which it can live, and then synchronized from the widgets back to the "working copy."

The workaround, to make that ergonomic, was to make it possible to "bind" UI widgets to target values -- for example, bind an individual DKFormPicker to a `form_reference_t&` in loaded form data; bind an individual QCheckBox to a particular bit in a particular `uint32_t&` flags mask; and so on. So for state that can live in a single widget, we bind the "working copy" field to that widget: the value is copied into the widget at bind time, and then we never pay attention to the value that lives in the "working copy;" and then when the widget's value is changed, the widget auto-updates the value that lives in the "working copy." And of course, these widgets are programmed to hook signals on `DovahKitCore`, detect when forms are deleted, and update their own internal state (which also triggers synchronization to the bound field in the "working copy"), thereby avoiding any dangling references.

For more complex state that can't live in a single widget, the typical approach is to manually copy that state out of the "working copy" and into someplace in the UI -- a private property on the dialog, the data of a QAbstractItemModel subclass, et cetera. We then copy that state back into the "working copy" when the dialog is closed. As long as these models also hook signals on `DovahKitCore`, they can sever any uses of to-be-deleted forms within their own internal contents; the data left behind in the "working copy" will dangle, but nothing should ever touch that.

But these factors lead to some pitfalls:

* Since the overwhelming majority of form-data that the UI needs to work with is individual fields, bound to perpetually-available single-purpose UI widgets, 90% of cases where we need to sever uses of a to-be-deleted form are handled automatically by the widgets themselves.

* Therefore, for the other 10% of cases, it's easy to forget to do the severing oneself, within the UI code.

* When you have fields on a form that the UI should never touch (like the event conditions on QUST, mentioned above), the UI has to go out of its way, at save time, to transplant those fields from the "canonical" form-data into the "working copy" form-data, so that when the latter overwrites the former, the fields are not clobbered by stale or potentially dangling data.

Thus, this commit: we are here fixing cases of reusable UI widgets, bespoke UI models, and form-editing dialogs failing to sever uses of to-be-deleted forms. This commit also modifies the QUST UI to avoid clobbering quest event conditions (as just described). Specific cases:

 - Debris: The model for Debris variants didn't sever uses of to-be-deleted TextureSets (by variants' TESModelTextureSwaps).

 - Faction: The list of interfaction relations allowed duplicate entries (i.e. multiple relationships with the same faction) in order to represent loaded form-data faithfully. However, the code for severing uses of a to-be-deleted faction would only remove the first-seen relationship with that faction.

 - Footstep Set: We just completely failed to sever uses at all. Someone call the Reno Sheriff's Department and let 'em know we found their missing lieutenant.

 - Music Track: Same mistake as Footstep Set.

 - Perk: The model for Entry Points had code to try and sever uses of to-be-deleted forms, but this code was written incorrectly -- a dumb mistake -- and thus didn't sever anything.

 - Quest: NOT YET FIXED, besides not clobbering event conditions. The code for quest stages, objectives, etc. still needs to be reviewed and proofed. Given how badly that old code sucks, I may well end up rewriting it entirely.

 - Scene: The `SceneFormVisualEditor` now severs uses of to-be-deleted Topics and Packages by Dialogue Actions and Package Actions. We also avoid dangling pointers to the owning quest or scene, though it'd be highly unusual for those to be deleted before the widget itself is disposed of.

 - `DKFormNIFPicker`: The widget, and its built-in edit dialog, will now properly sever uses of to-be-deleted TextureSets.

----

Aside from all of the above, this commit also synchronizes the `widgets` folder in the Qt Designer plug-in project with that of the main project. This means that several widgets that aren't actually designable are now in the plug-in's source code. I have bigger things to worry about, so it's staying.

+++ native/src/DovahKitQtCustomWidgets/widgets/DKFormDestructionDataButton.h
@@ -26,7 +26,7 @@ class DKFormDestructionDataButton : public QWidget {
          Q_DECLARE_FLAGS(DestructionStageFlags, DestructionStageFlag);
          
          struct DestructionStage {
-            uint16_t                health_percent   = 100;
+            uint8_t                 health_percent   = 100;
             uint8_t                 damage_stage     = 0;
             DestructionStageFlags   flags            = {};
             uint32_t                self_damage_rate = 0; // the object inflicts this much damage on itself per second (i.e. health decay)
